split inspect into list and info
